backend/requirements.txt
Flask
Flask_Cors
groq
numpy
pandas
pypdf
python-dotenv
scikit_learn
sentence_transformers
Werkzeug
